pyproject.toml
[tool.poetry]
name = "dash-competencias"
version = "0.1.0"
description = "Dashboard para consulta de Competencias Digitales Docentes de los cursos de Aualragón"
authors = ["deleyva <jesuslopezdeleyva@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
dash = "^2.7.0"
pandas = "^1.5.1"
jupyter = "^1.0.0"
jupyter-dash = "^0.4.2"
Pillow = "^9.3.0"
python-slugify = "^6.1.2"
dash-bootstrap-components = "^1.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
